Fix what 'larger' actually means
Fixes #63

Previously we were interpreting 'larger' as "which equation has a higher
y-value?".  This is actually incorrect if we're rotating above the two
equations.
The real interpretation should be "which equation is farther away from
the axis of rotation?".
This PR fixes that and thankfully also untangles the ugly double-return
that getIntersectionWith() was doing.
In addition, a possible bug with boundY2 was fixed (though I'm still not
entirely sure I know what that's doing).

Test cases all pass.